requirements.txt
anyio==3.6.2
certifi==2023.5.7
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.2
h11==0.14.0
httpcore==0.17.1
httptools==0.5.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
orjson==3.8.10
pyaes==1.6.1
pyasn1==0.5.0
pydantic==1.10.7
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
rsa==4.9
sniffio==1.3.0
starlette==0.27.0
Telethon==1.28.5
typing-extensions==4.5.0
ujson==5.7.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
